organise test files
